+++ package.json
@@ -4,8 +4,8 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "server-dev": "nodemon server/index.js",
-    "start": "node server/index.js",
+    "server-dev": "nodemon index.js",
+    "start": "node index.js",
     "test": "jest --detectOpenHandles --coverage"
   },
   "repository": {
