+++ sequelize/models/characteristic.model.js
