+++ sequelize/models/characteristic_review.model.js
